Split DB and NOAA api calls, create initial hourly calls
