+++ src/app/Contexts/UserContext.tsx
@@ -3,7 +3,7 @@ import React, { createContext, useState, useEffect } from "react";
 import { getSession } from "next-auth/react";
 import { UserSessionInfo } from "../Interfaces/interfaces";
 import { UserLocation } from "../Classes/UserLocation";
-import { getAllUserLocations } from "../Util/APICalls";
+import { getAllUserLocations } from "../Util/DatabaseApiCalls";
 
 interface UserContextType {
   userInfo: UserSessionInfo | null | undefined;
